put whatever composable functionality was needed into the components to help reduce clutter to see what i did and did not need
