+++ frontend/src/components/fly/FlyList.vue
@@ -1,5 +1,6 @@
 <template>
-  <div v-for="fly in flys" :key="fly.id">
+  <fly-search v-model="searchTerm"></fly-search>
+  <div v-for="fly in filteredFlys" :key="fly.id">
     <h3>
       <router-link :to="{ name: 'Fly', params: { id: fly.id } }">{{
         fly.pattern.name
@@ -9,15 +10,23 @@
 </template>
 
 <script>
-import { ref } from "vue";
+import { ref, computed } from "vue";
+import FlySearch from "@/components/fly/FlySearch.vue";
 
 export default {
+  components: { FlySearch },
   props: ["list"],
   setup(props) {
     const flys = ref(props.list);
+    const searchTerm = ref('');
+
+    const filteredFlys = computed(() => {
+      return flys.value.filter((f) => f.name.toLowerCase().includes(searchTerm.value));
+    })
 
     return {
-      flys,
+      filteredFlys,
+      searchTerm
     };
   },
 };
