+++ frontend/src/components/fly/FlySearch.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import {computed} from 'vue';
+import { computed } from 'vue';
 
 export default {
     emits: ['update:modelValue'],
